refactor: 문제 풀기에 있던 examresult 모듈화
