+++ feature-ui-exam-result/src/main/java/team/duckie/app/android/feature/ui/exam/result/viewmodel/ExamResultSideEffect.kt
@@ -1,0 +1,14 @@
+/*
+ * Designed and developed by Duckie Team, 2022
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
+ */
+
+package team.duckie.app.android.feature.ui.exam.result.viewmodel
+
+sealed class ExamResultSideEffect {
+    object FinishExamResult : ExamResultSideEffect()
+    data class SendToast(val message: String) : ExamResultSideEffect()
+    data class ReportError(val exception: Throwable) : ExamResultSideEffect()
+}
