+++ feature-ui-exam-result/src/main/java/team/duckie/app/android/feature/ui/exam/result/viewmodel/ExamResultState.kt
@@ -1,0 +1,13 @@
+/*
+ * Designed and developed by Duckie Team, 2022
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
+ */
+
+package team.duckie.app.android.feature.ui.exam.result.viewmodel
+
+data class ExamResultState(
+    val isReportLoading: Boolean = true,
+    val reportUrl: String = "",
+)
