--- feature-ui-solve-problem/src/main/kotlin/team/duckie/app/android/feature/ui/solve/problem/examresult/viewmodel/ExamResultState.kt
@@ -1,13 +1,0 @@
-/*
- * Designed and developed by Duckie Team, 2022
- *
- * Licensed under the MIT.
- * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
- */
-
-package team.duckie.app.android.feature.ui.solve.problem.examresult.viewmodel
-
-data class ExamResultState(
-    val isReportLoading: Boolean = true,
-    val reportUrl: String = "",
-)
